+++ etc/cloud-build-push-charts.sh
@@ -1,14 +1,21 @@
 #!/usr/bin/env bash
-# Push our charts up to a gs storage bucket for Helm
-# You can add this as a Helm repo
-# updated until the second pass.   This should be split into two phases - the first for base charts,
-# the second for cmp-* charts.
-cdir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+# Push our charts up to a gs storage bucket for Helm.
+# Runs as a cloud build step, so we need to download helm.
 
 BUCKET=forgerock-charts
 URL="https://storage.googleapis.com/forgerock-charts"
 
-hdir=${cdir}/../helm
+# Where our helm charts are located.
+hdir=`pwd`/helm
+
+
+# The previous build step downloaded Helm to our working directory. We need to unpack it.
+tar xvf helm.tar.gz
+
+helm=`pwd`/linux-amd64/helm
+chmod +x $helm
+
+$helm init --client-only
 
 dir=/tmp/charts
 
@@ -19,18 +26,21 @@ charts="opendj amster openam openidm openig postgres-openidm frcommon git opendj
 for chart in $charts
 do
     echo "Packaging $chart"
-    helm dep update $hdir/$chart
-    helm package $hdir/$chart
+    $helm dep update $hdir/$chart
+    $helm package $hdir/$chart
 done
 
+# Fetch a copy of the existing index.
 gsutil cp gs://${BUCKET}/index.yaml .
-helm repo index --url $URL --merge index.yaml .
+# Merge the new charts with the existing index.
+$helm repo index --url $URL --merge index.yaml .
 
+# Copy all the charts and index up to our bucket.
 gsutil -m rsync ./ gs://${BUCKET}
 
+# Make the charts world readable.
 gsutil -m acl set -R -a public-read gs://${BUCKET}
 
 # See https://github.com/kubernetes/helm/issues/2453.
 # This makes sure the bucket is not cached (default is to cache https:// objects for 1 hour).
 gsutil -m setmeta -h "Content-Type:text/html" -h "Cache-Control:private, max-age=0, no-transform" "gs://${BUCKET}/*"
-